+++ sematic/cli/BUILD
@@ -33,7 +33,7 @@ sematic_py_lib(
     ],
     deps = [
         "//sematic/db:migrate_lib",
-        "//sematic:logging",
+        "//sematic:logs",
     ],
 )
 
@@ -58,7 +58,7 @@ sematic_py_lib(
     deps = [
         ":cli",
         "//sematic:api_client",
-        "//sematic:logging",
+        "//sematic:logs",
         "//sematic/api/endpoints:auth",
         "//sematic/config:user_settings",
         "//sematic/config:config",
